[test] [add] message index to check message format
